requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
django-ratelimit==4.1.0
django-axes==6.1.1
celery==5.3.4